+++ apps/models/form/Auth/UserConfirmForm.php
@@ -105,7 +105,7 @@ class UserConfirmForm extends Validator
         $this->update_confirm_status();
 
         // Send user email to tell his new password.
-        Site::sendEmail([$this->email], 'New Password',
+        app()->site->sendEmail([$this->email], 'New Password',
             'email/user_new_password', [
                 'username' => $this->username,
                 'password' => $new_password,