+++ apps/models/form/UserInviteForm.php
@@ -152,7 +152,7 @@ class UserInviteForm extends UserRegisterForm
         }
 
         if ($invite_status === true) { // TODO use email queue
-            Site::sendEmail([$this->email], 'Invite To ' . config('base.site_name'),
+            app()->site->sendEmail([$this->email], 'Invite To ' . config('base.site_name'),
                 'email/user_invite', [
                     'username' => $this->username,
                     'invite_link' => $this->invite_link,